requirements.txt
aiohttp
aiosignal
async-timeout
attrs
certifi
charset-normalizer
colorama
et-xmlfile
frozenlist
idna
multidict
numpy
openai
openpyxl
pandas
pandas-stubs
praw
psycopg2-binary
prawcore
py-cord
python-dateutil
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0


python-dotenv
pytz
requests
six
tqdm
types-pytz
typing_extensions
update-checker
urllib3
websocket-client
yarl
